+++ cli/operator/keys/create_test.go
@@ -3,6 +3,7 @@ package keys
 import (
 	"fmt"
 	"os"
+	"path/filepath"
 	"testing"
 
 	prompterMock "github.com/NethermindEth/eigenlayer/cli/prompter/mocks"
@@ -11,6 +12,11 @@ import (
 )
 
 func TestCreateCmd(t *testing.T) {
+	homePath, err := os.UserHomeDir()
+	if err != nil {
+		t.Fatal(err)
+	}
+
 	tests := []struct {
 		name       string
 		args       []string
@@ -66,7 +72,7 @@ func TestCreateCmd(t *testing.T) {
 			promptMock: func(p *prompterMock.MockPrompter) {
 				p.EXPECT().InputHiddenString(gomock.Any(), gomock.Any(), gomock.Any()).Return("", nil)
 			},
-			keyPath: OperatorKeyFolder + "/test.ecdsa.key.json",
+			keyPath: filepath.Join(homePath, OperatorKeystoreSubFolder, "/test.ecdsa.key.json"),
 		},
 		{
 			name: "valid bls key creation",
@@ -75,13 +81,13 @@ func TestCreateCmd(t *testing.T) {
 			promptMock: func(p *prompterMock.MockPrompter) {
 				p.EXPECT().InputHiddenString(gomock.Any(), gomock.Any(), gomock.Any()).Return("", nil)
 			},
-			keyPath: OperatorKeyFolder + "/test.bls.key.json",
+			keyPath: filepath.Join(homePath, OperatorKeystoreSubFolder, "/test.bls.key.json"),
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Cleanup(func() {
-				_ = os.RemoveAll(OperatorKeyFolder)
+				_ = os.Remove(tt.keyPath)
 			})
 			controller := gomock.NewController(t)
 			p := prompterMock.NewMockPrompter(controller)