+++ cli/operator/keys/import_test.go
@@ -4,6 +4,7 @@ import (
 	"encoding/hex"
 	"fmt"
 	"os"
+	"path/filepath"
 	"strings"
 	"testing"
 
@@ -15,6 +16,11 @@ import (
 )
 
 func TestImportCmd(t *testing.T) {
+	homePath, err := os.UserHomeDir()
+	if err != nil {
+		t.Fatal(err)
+	}
+
 	tests := []struct {
 		name       string
 		args       []string
@@ -86,7 +92,7 @@ func TestImportCmd(t *testing.T) {
 			promptMock: func(p *prompterMock.MockPrompter) {
 				p.EXPECT().InputHiddenString(gomock.Any(), gomock.Any(), gomock.Any()).Return("", nil)
 			},
-			keyPath: OperatorKeyFolder + "/test.ecdsa.key.json",
+			keyPath: filepath.Join(homePath, OperatorKeystoreSubFolder, "/test.ecdsa.key.json"),
 		},
 		{
 			name: "valid ecdsa key import with 0x prefix",
@@ -95,7 +101,7 @@ func TestImportCmd(t *testing.T) {
 			promptMock: func(p *prompterMock.MockPrompter) {
 				p.EXPECT().InputHiddenString(gomock.Any(), gomock.Any(), gomock.Any()).Return("", nil)
 			},
-			keyPath: OperatorKeyFolder + "/test.ecdsa.key.json",
+			keyPath: filepath.Join(homePath, OperatorKeystoreSubFolder, "/test.ecdsa.key.json"),
 		},
 		{
 			name: "valid bls key import",
@@ -104,13 +110,13 @@ func TestImportCmd(t *testing.T) {
 			promptMock: func(p *prompterMock.MockPrompter) {
 				p.EXPECT().InputHiddenString(gomock.Any(), gomock.Any(), gomock.Any()).Return("", nil)
 			},
-			keyPath: OperatorKeyFolder + "/test.bls.key.json",
+			keyPath: filepath.Join(homePath, OperatorKeystoreSubFolder, "/test.bls.key.json"),
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Cleanup(func() {
-				_ = os.RemoveAll(OperatorKeyFolder)
+				_ = os.Remove(tt.keyPath)
 			})
 			controller := gomock.NewController(t)
 			p := prompterMock.NewMockPrompter(controller)